Refactored HomePage & Routing
