+++ package.json
@@ -4,12 +4,15 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --host",
     "build": "tsc -b && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "preact": "^10.23.1"
+    "flowbite": "^2.5.1",
+    "keen-slider": "^6.8.6",
+    "preact": "^10.23.2",
+    "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
     "@preact/preset-vite": "^2.9.0",
