--- src/PageTitle.tsx
@@ -1,15 +1,0 @@
-export default function PageTitle() {
-  return (
-    <>
-      <div className="flex justify-center mb-4 text-4xl leading-none tracking-tight text-gray-900 md:text-5xl lg:text-6xl dark:text-white">
-        <h1>Powerful, Versatile. </h1>
-      </div>
-      <div className="flex justify-center mb-4 text-4xl leading-none tracking-tight text-[#444dee] md:text-5xl lg:text-6xl ">
-        <h1>PacketEvents</h1>
-      </div>
-      <div className="flex justify-center mb-6 text-lg font-normal text-gray-500 lg:text-xl sm:px-16 xl:px-48 dark:text-gray-400">
-        <h2>Achieve next-level networking by using PacketEvents.</h2>
-      </div>
-    </>
-  );
-}
